ログインフィルター、checkAdmin()を再適用, 日報詳細画面を編集

+++ src/main/java/actions/EmployeeAction.java
@@ -43,7 +43,7 @@ public class EmployeeAction extends ActionBase {
     public void index() throws ServletException, IOException {
 
         //管理者かどうかのチェック
-//        if (checkAdmin()) {
+        if (checkAdmin()) {
             //指定されたページ数の一覧画面に表示するデータを取得
             int page = getPage();
             List<EmployeeView> employees = service.getPerPage(page);
@@ -66,7 +66,7 @@ public class EmployeeAction extends ActionBase {
             //一覧画面を表示
             forward(ForwardConst.FW_EMP_INDEX);
         }
-//    }
+    }
 
     /**
      * 新規登録画面を表示する
@@ -94,7 +94,7 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkToken() && checkAdmin()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -146,7 +146,7 @@ public class EmployeeAction extends ActionBase {
     public void show() throws ServletException, IOException {
 
         //管理者かどうかのチェック
-//        if (checkAdmin()) {
+        if (checkAdmin()) {
 
             //idを条件に従業員データを取得する
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -163,7 +163,7 @@ public class EmployeeAction extends ActionBase {
             //詳細画面を表示
             forward(ForwardConst.FW_EMP_SHOW);
         }
-//    }
+    }
 
     /**
      * 編集画面を表示する
@@ -173,7 +173,7 @@ public class EmployeeAction extends ActionBase {
     public void edit() throws ServletException, IOException {
 
         //管理者かどうかのチェック
-//        if (checkAdmin()) {
+        if (checkAdmin()) {
 
             //idを条件に従業員データを取得する
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -191,7 +191,7 @@ public class EmployeeAction extends ActionBase {
             //編集画面を表示する
             forward(ForwardConst.FW_EMP_EDIT);
         }
-//    }
+    }
 
     /**
      * 更新を行う
@@ -201,7 +201,7 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkToken() && checkAdmin()) {
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -251,7 +251,7 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkToken() && checkAdmin()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
